feat: refactor authentication components and implement new forms
- Replaced `LoginForm` and `RegisterForm` with `LoginView` and `RegisterView` components for better structure.
- Introduced new `LoginForm` and `RegisterForm` components with enhanced functionality and validation.
- Updated hooks for authentication to improve user experience and error handling.
- Adjusted imports to reflect the new component structure.
