+++ app/(auth)/login/page.tsx
@@ -1,5 +1,5 @@
-import LoginForm from "./LoginForm";
+import LoginView from "@/features/auth/views/LoginView";
 
 export default async function Login() {
-  return <LoginForm />;
+  return <LoginView />;
 }
