+++ app/(auth)/register/page.tsx
@@ -1,5 +1,5 @@
-import RegisterForm from "./RegisterForm";
+import RegisterView from "@/features/auth/views/RegisterView";
 
 export default async function page() {
-  return <RegisterForm />;
+  return <RegisterView />;
 }
