+++ features/auth/hooks/useAuth.tsx
@@ -3,7 +3,7 @@ import {
   loginWithCredentials,
   loginWithProvider,
   registerUser,
-} from "@/lib/api/authApi";
+} from "@/features/auth/lib/authApi";
 import { LoginFormData, RegisterFormData } from "@/schemas/auth.schema";
 import { useMutation, useQuery } from "@tanstack/react-query";
 import { useRouter } from "next/navigation";
