+++ features/auth/views/LoginView.tsx
@@ -1,0 +1,5 @@
+import LoginForm from "../components/LoginForm";
+
+export default function LoginView() {
+  return <LoginForm />;
+}
