+++ features/auth/views/RegisterView.tsx
@@ -1,0 +1,5 @@
+import RegisterForm from "../components/RegisterForm";
+
+export default function RegisterView() {
+  return <RegisterForm />;
+}
